requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
pydantic>=1.10.0
transformers>=4.28.0
torch>=2.0.0
ibm-watsonx-ai>=1.0.0